+++ CHANGELOG.md
@@ -42,6 +42,9 @@ All notable changes for each version of this project will be documented in this 
 - `IgxGridState`
     -  When possible the state directive nows reuses the column that already exists on the grid when restoring the state, instead of creating new column instances every time. This removes the need to set any complex objects manually back on the column on `columnInit`. The only instance where this is still necessary is when the column (or its children in case of column groups) have no `field` property so there's no way to uniquely identify the matching column.
     - Added support for persisting Multi-Row Layout.
+- `IgxBanner`
+    - Added a new `collapsed` input property, enabling control over the initial state of the banner, allowing it to be either collapsed (hidden) or expanded (visible) upon rendering.
+
 ### Themes
 - `Palettes`
     - All palette colors have been migrated to the [CSS relative colors syntax](https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_colors/Relative_colors). This means that color consumed as CSS variables no longer need to be wrapped in an `hsl` function.